+++ src/engine/strat_engine/keys.rs
@@ -2,25 +2,16 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-use std::{
-    convert::TryFrom,
-    ffi::CString,
-    fs::File,
-    io::{self, Read},
-    mem::size_of,
-    os::unix::io::{FromRawFd, RawFd},
-    str,
-};
+use std::{convert::TryFrom, ffi::CString, io, mem::size_of, os::unix::io::RawFd, str};
 
 use libc::{syscall, SYS_add_key, SYS_keyctl};
-use termios::Termios;
 
-use devicemapper::Bytes;
 use libcryptsetup_rs::SafeMemHandle;
 
 use crate::{
     engine::{
         engine::{KeyActions, MAX_STRATIS_PASS_SIZE},
+        shared,
         strat_engine::names::KeyDescription,
         types::{DeleteAction, KeySerial, MappingCreateAction, SizedKeyMemory},
     },
@@ -385,69 +376,11 @@ impl KeyActions for StratKeyActions {
         key_fd: RawFd,
         interactive: Option<bool>,
     ) -> StratisResult<MappingCreateAction<()>> {
-        fn read_loop(
-            bytes_iter: &mut io::Bytes<File>,
-            mem: &mut [u8],
-            interactive: bool,
-        ) -> StratisResult<usize> {
-            let mut pos = 0;
-            while pos < MAX_STRATIS_PASS_SIZE {
-                match bytes_iter.next() {
-                    Some(Ok(b)) => {
-                        if interactive && b as char == '\n' {
-                            break;
-                        }
-
-                        mem[pos] = b;
-                        pos += 1;
-                    }
-                    Some(Err(e)) => return Err(e.into()),
-                    None => break,
-                }
-            }
-            Ok(pos)
-        }
-
-        let key_file = unsafe { File::from_raw_fd(key_fd) };
-        let mut memory = SafeMemHandle::alloc(MAX_STRATIS_PASS_SIZE)?;
-
-        let old_attrs = if let Some(true) = interactive {
-            let old_attrs = Termios::from_fd(key_fd)?;
-            let mut new_attrs = old_attrs;
-            new_attrs.c_lflag &= !(termios::ICANON | termios::ECHO);
-            new_attrs.c_cc[termios::VMIN] = 1;
-            new_attrs.c_cc[termios::VTIME] = 0;
-            termios::tcsetattr(key_fd, termios::TCSANOW, &new_attrs)?;
-            Some(old_attrs)
-        } else {
-            None
-        };
-
-        let mut bytes_iter = key_file.bytes();
-
-        let res = read_loop(&mut bytes_iter, memory.as_mut(), interactive.is_some());
-
-        if let Some(ref oa) = old_attrs {
-            termios::tcsetattr(key_fd, termios::TCSANOW, oa)?;
-        }
-
-        let pos = res?;
-
-        if pos == MAX_STRATIS_PASS_SIZE && bytes_iter.next().is_some() {
-            return Err(StratisError::Engine(
-                ErrorEnum::Invalid,
-                format!(
-                    "Provided key exceeded maximum allow length of {}",
-                    Bytes(MAX_STRATIS_PASS_SIZE as u64)
-                ),
-            ));
-        }
-
-        let sized_memory = SizedKeyMemory::new(memory, pos);
+        let memory = shared::set_key_shared(key_fd, interactive)?;
 
         Ok(set_key_idem(
             &KeyDescription::try_from(key_desc.to_string())?,
-            sized_memory,
+            memory,
         )?)
     }
